Error handling and dead code deletion

+++ backend/controller/auth.go
@@ -6,20 +6,38 @@ import (
 	"api/helpers"
 	"api/models"
 	"encoding/json"
+	"github.com/go-chi/chi/v5"
 	webauthn2 "github.com/go-webauthn/webauthn/webauthn"
 	"net/http"
+	"strconv"
 )
 
 // https://developers.google.com/codelabs/passkey-form-autofill
 
+func getUser(w http.ResponseWriter, r *http.Request) (user models.User, valid bool) {
+	valid = true
+	id, err := strconv.ParseUint(chi.URLParam(r, "userId"), 10, 32)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusBadRequest)
+		valid = false
+		return
+	}
+
+	result := db.DB.First(&user, id)
+	if result.Error != nil {
+		helpers.DBErrorHandling(result.Error, w)
+		valid = false
+	}
+
+	return
+}
+
 // https://webauthn.guide/#registration
 func RegisterRequest(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
-	var user models.User
-	result := db.DB.First(&user) // TODO: parse user from request body
 
-	if result.Error != nil {
-		helpers.DBErrorHandling(result.Error, w)
+	user, valid := getUser(w, r)
+	if !valid {
 		return
 	}
 
@@ -37,11 +55,12 @@ func RegisterRequest(w http.ResponseWriter, r *http.Request) {
 
 	// handle errors
 	if err != nil {
-		// TODO: handle error
+		http.Error(w, "Could not begin registration for provided user", http.StatusInternalServerError)
+		return
 	}
 
 	// store sesion data
-	result = db.DB.Create(&sessionData)
+	result := db.DB.Create(&sessionData)
 	if result.Error != nil {
 		helpers.DBErrorHandling(result.Error, w)
 		return
@@ -49,7 +68,7 @@ func RegisterRequest(w http.ResponseWriter, r *http.Request) {
 
 	errJson := json.NewEncoder(w).Encode(&options.Response)
 	if errJson != nil {
-		http.Error(w, "500", http.StatusInternalServerError)
+		http.Error(w, "Could not return results", http.StatusInternalServerError)
 		return
 	}
 }
@@ -57,11 +76,8 @@ func RegisterRequest(w http.ResponseWriter, r *http.Request) {
 func RegisterResponse(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 
-	// Get the user
-	var user models.User
-	result := db.DB.First(&user) // TODO: parse user from request
-	if result.Error != nil {
-		helpers.DBErrorHandling(result.Error, w)
+	user, valid := getUser(w, r)
+	if !valid {
 		return
 	}
 
@@ -69,8 +85,10 @@ func RegisterResponse(w http.ResponseWriter, r *http.Request) {
 	sessionData := models.SessionData{
 		UserId: user.WebAuthnID(),
 	}
-	result = db.DB.First(&sessionData)
-	db.DB.Delete(&sessionData)
+	result := db.DB.First(&sessionData)
+	if result.Error == nil {
+		result = db.DB.Delete(&sessionData)
+	}
 	if result.Error != nil {
 		helpers.DBErrorHandling(result.Error, w)
 		return
@@ -87,8 +105,7 @@ func RegisterResponse(w http.ResponseWriter, r *http.Request) {
 
 	credential, err := config.WebAuthn.FinishRegistration(user, session, r)
 	if err != nil {
-		// TODO: Handle Error
-
+		http.Error(w, "Could not finish registration for provided user", http.StatusInternalServerError)
 		return
 	}
 
@@ -96,20 +113,6 @@ func RegisterResponse(w http.ResponseWriter, r *http.Request) {
 	transports := make([]string, 0)
 	for _, val := range credential.Transport {
 		transports = append(transports, string(val))
-	}
-
-	flags := models.Flags{
-		UserPresent:    credential.Flags.UserPresent,
-		UserVerified:   credential.Flags.UserVerified,
-		BackupEligible: credential.Flags.BackupEligible,
-		BackupState:    credential.Flags.BackupState,
-	}
-
-	authenticator := models.Authenticator{
-		AAGUID:       credential.Authenticator.AAGUID,
-		SignCount:    credential.Authenticator.SignCount,
-		CloneWarning: credential.Authenticator.CloneWarning,
-		Attachment:   string(credential.Authenticator.Attachment),
 	}
 
 	// Add public key to DB
@@ -120,9 +123,20 @@ func RegisterResponse(w http.ResponseWriter, r *http.Request) {
 		PublicKey:       credential.PublicKey,
 		AttestationType: credential.AttestationType,
 		Transport:       transports,
-		Flags:           flags,
-		Authenticator:   authenticator,
+		Flags: models.Flags{
+			UserPresent:    credential.Flags.UserPresent,
+			UserVerified:   credential.Flags.UserVerified,
+			BackupEligible: credential.Flags.BackupEligible,
+			BackupState:    credential.Flags.BackupState,
+		},
+		Authenticator: models.Authenticator{
+			AAGUID:       credential.Authenticator.AAGUID,
+			SignCount:    credential.Authenticator.SignCount,
+			CloneWarning: credential.Authenticator.CloneWarning,
+			Attachment:   string(credential.Authenticator.Attachment),
+		},
 	}
+
 	result = db.DB.Create(&dbCredential)
 	if result.Error != nil {
 		helpers.DBErrorHandling(result.Error, w)
@@ -133,7 +147,7 @@ func RegisterResponse(w http.ResponseWriter, r *http.Request) {
 
 	errJson := json.NewEncoder(w).Encode("Registration Success")
 	if errJson != nil {
-		http.Error(w, "500", http.StatusInternalServerError)
+		http.Error(w, "Could not return results", http.StatusInternalServerError)
 		return
 	}
 }
@@ -142,22 +156,14 @@ func SigninRequest(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 
 	// Find the user
-	body := struct {
-		Id uint `json:"id"`
-	}{}
-	err := json.NewDecoder(r.Body).Decode(&body)
-
-	var user models.User
-	result := db.DB.First(&user, body.Id)
-	if result.Error != nil {
-		helpers.DBErrorHandling(result.Error, w)
+	user, valid := getUser(w, r)
+	if !valid {
 		return
 	}
 
 	options, session, err := config.WebAuthn.BeginLogin(user)
 	if err != nil {
-		// Handle Error and return.
-
+		http.Error(w, "Could not being log in", http.StatusInternalServerError)
 		return
 	}
 
@@ -171,13 +177,8 @@ func SigninRequest(w http.ResponseWriter, r *http.Request) {
 		Extensions:           session.Extensions,
 	}
 
-	// handle errors
-	if err != nil {
-		// TODO: handle error
-	}
-
 	// store sesion data
-	result = db.DB.Create(&sessionData)
+	result := db.DB.Create(&sessionData)
 	if result.Error != nil {
 		helpers.DBErrorHandling(result.Error, w)
 		return
@@ -185,7 +186,7 @@ func SigninRequest(w http.ResponseWriter, r *http.Request) {
 
 	errJson := json.NewEncoder(w).Encode(&options.Response)
 	if errJson != nil {
-		http.Error(w, "500", http.StatusInternalServerError)
+		http.Error(w, "Could not return results", http.StatusInternalServerError)
 		return
 	}
 }
@@ -194,10 +195,8 @@ func SigninResponse(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 
 	// Get the user
-	var user models.User
-	result := db.DB.First(&user) // TODO: parse user from request
-	if result.Error != nil {
-		helpers.DBErrorHandling(result.Error, w)
+	user, valid := getUser(w, r)
+	if !valid {
 		return
 	}
 
@@ -205,8 +204,10 @@ func SigninResponse(w http.ResponseWriter, r *http.Request) {
 	sessionData := models.SessionData{
 		UserId: user.WebAuthnID(),
 	}
-	result = db.DB.First(&sessionData)
-	db.DB.Delete(&sessionData)
+	result := db.DB.First(&sessionData)
+	if result.Error == nil {
+		result = db.DB.Delete(&sessionData)
+	}
 	if result.Error != nil {
 		helpers.DBErrorHandling(result.Error, w)
 		return
@@ -223,8 +224,7 @@ func SigninResponse(w http.ResponseWriter, r *http.Request) {
 
 	credential, err := config.WebAuthn.FinishLogin(user, session, r)
 	if err != nil {
-		// TODO: Handle Error
-
+		http.Error(w, "Could not finish log in", http.StatusInternalServerError)
 		return
 	}
 
@@ -234,28 +234,25 @@ func SigninResponse(w http.ResponseWriter, r *http.Request) {
 		transports = append(transports, string(val))
 	}
 
-	flags := models.Flags{
-		credential.Flags.UserPresent,
-		credential.Flags.UserVerified,
-		credential.Flags.BackupEligible,
-		credential.Flags.BackupState,
-	}
-
-	authenticator := models.Authenticator{
-		AAGUID:       credential.Authenticator.AAGUID,
-		SignCount:    credential.Authenticator.SignCount,
-		CloneWarning: credential.Authenticator.CloneWarning,
-		Attachment:   string(credential.Authenticator.Attachment),
-	}
-
 	dbCredential := models.Credential{
 		WebauthnId:      credential.ID,
 		PublicKey:       credential.PublicKey,
 		AttestationType: credential.AttestationType,
 		Transport:       transports,
-		Flags:           flags,
-		Authenticator:   authenticator,
+		Flags: models.Flags{
+			UserPresent:    credential.Flags.UserPresent,
+			UserVerified:   credential.Flags.UserVerified,
+			BackupEligible: credential.Flags.BackupEligible,
+			BackupState:    credential.Flags.BackupState,
+		},
+		Authenticator: models.Authenticator{
+			AAGUID:       credential.Authenticator.AAGUID,
+			SignCount:    credential.Authenticator.SignCount,
+			CloneWarning: credential.Authenticator.CloneWarning,
+			Attachment:   string(credential.Authenticator.Attachment),
+		},
 	}
+
 	result = db.DB.Save(&dbCredential)
 	if result.Error != nil {
 		helpers.DBErrorHandling(result.Error, w)
@@ -264,7 +261,7 @@ func SigninResponse(w http.ResponseWriter, r *http.Request) {
 
 	errJson := json.NewEncoder(w).Encode("Login Success")
 	if errJson != nil {
-		http.Error(w, "500", http.StatusInternalServerError)
+		http.Error(w, "Could not return results", http.StatusInternalServerError)
 		return
 	}
 }
